+++ src/arena.rs
@@ -8,7 +8,7 @@ use alloc::vec::Vec;
 
 use crate::free_pointer::FreePointer;
 use crate::generation::Generation;
-use crate::iter::{Drain, IntoIter, Iter, IterMut};
+use crate::iter::{Drain, IntoIter, Iter, IterMut, Values};
 
 /// Container that can have elements inserted into it and removed from it.
 ///
@@ -575,6 +575,13 @@ impl<T> Arena<T> {
             inner: self.storage.iter().enumerate(),
             len: self.len,
         }
+    }
+
+    /// Iterate over all of the values contained in the arena.
+    ///
+    /// Iteration order is not defined.
+    pub fn values(&self) -> Values<'_, T> {
+        Values { inner: self.iter() }
     }
 
     /// Iterate over all of the indexes and values contained in the arena, with