+++ src/iter/mod.rs
@@ -4,8 +4,10 @@ mod drain;
 mod into_iter;
 mod iter;
 mod iter_mut;
+mod values;
 
 pub use drain::Drain;
 pub use into_iter::IntoIter;
 pub use iter::Iter;
 pub use iter_mut::IterMut;
+pub use values::Values;